refactor(api): rename getExternalAuthenticationRedirectionUrl usecase
to getSamlAuthenticationRedirectionUrl
