+++ api/src/authentication/domain/usecases/index.js
@@ -1,6 +1,8 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
+import * as authenticationMethodRepository from '../../../../lib/infrastructure/repositories/authentication-method-repository.js';
+import { config } from '../../../shared/config.js';
 import { tokenService } from '../../../shared/domain/services/token-service.js';
 import * as adminMemberRepository from '../../../shared/infrastructure/repositories/admin-member-repository.js';
 import * as userLoginRepository from '../../../shared/infrastructure/repositories/user-login-repository.js';
@@ -14,6 +16,8 @@ import * as refreshTokenService from '../services/refresh-token-service.js';
 const path = dirname(fileURLToPath(import.meta.url));
 
 const dependencies = {
+  authenticationMethodRepository,
+  config,
   refreshTokenService,
   pixAuthenticationService,
   tokenService,
